requirements.txt
Flask==2.0.2
SQLAlchemy==1.4.29
psycopg2-binary==2.9.3